Changed "stacks" to make it more clear.
In current game situation, there could be some questions.
If Bob gets "red" clue, it always red 1 for beginners. Because Alice could clue "2's", which would be a better save with single meaning. (Other 2's already played and Bob need only focused). May be it's a bit tricky for beginners, or it was a purpose that reader can make a mistake here.
In my opinion, if we change stacks to empty, it would better correspond to given solution.